requirements.txt
discord.py
aiohttp
aiofiles
requests
json
os
asyncio
random
datetime